+++ UnityProject/Assets/Prefabs/Items/Implants/HellSpawn/Tiefling/TieflingTongue.prefab
@@ -11,7 +11,7 @@ PrefabInstance:
     - target: {fileID: 3113739160240451436, guid: 57bdc54d41899c34a897a26ef0a10933,
         type: 3}
       propertyPath: speechModifiers
-      value: 8
+      value: 4104
       objectReference: {fileID: 0}
     - target: {fileID: 3113739160240451436, guid: 57bdc54d41899c34a897a26ef0a10933,
         type: 3}
