+++ UnityProject/Assets/Resources/ScriptableObjectsSingletons/SpeechModManager.asset
@@ -13,7 +13,7 @@ MonoBehaviour:
   m_Name: SpeechModManager
   m_EditorClassIdentifier: 
   speechModifier:
-    m_keys: 02000000000800000010000000200000004000000080000000000100000002000000040000000800100000000000200001000000
+    m_keys: 0200000000080000001000000020000000400000008000000000010000000200000004000000080010000000000020000100000008000000
     m_values:
     - {fileID: 11400000, guid: fa3adcd3d8a530f44acd8e18df0319bb, type: 2}
     - {fileID: 11400000, guid: 348e419ebc24a4b4eafdc8239b8aaedb, type: 2}
@@ -28,3 +28,4 @@ MonoBehaviour:
     - {fileID: 11400000, guid: f84283fd290c74946830fdd6c0cd3c04, type: 2}
     - {fileID: 11400000, guid: 9c9ccc21e0aee3b45a2463e8c42adc90, type: 2}
     - {fileID: 11400000, guid: 2691de9037e77db4eb9d0e94be5fe6c5, type: 2}
+    - {fileID: 11400000, guid: 772eef0e46246a146adb3c48809e2b4f, type: 2}
