+++ UnityProject/Assets/ScriptableObjects/Speech/CustomMods/Hiss.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5446a8d2ab4b0f246be6dfba2996eb90
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
