+++ UnityProject/Assets/ScriptableObjects/Speech/CustomMods/HissMod.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f8e3f2faebb244f29993ffd06effd675
+timeCreated: 1714021811
