+++ UnityProject/Assets/ScriptableObjects/Speech/Hiss.asset
@@ -1,0 +1,22 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c302747b295f59d429cc78a122bcb51d, type: 3}
+  m_Name: Hiss
+  m_EditorClassIdentifier: 
+  activateReplacements: 1
+  wordReplaceList: []
+  letterReplaceList: []
+  activateAdditions: 0
+  probability: 0
+  beginning: []
+  ending: []
+  customCode: {fileID: 11400000, guid: 5446a8d2ab4b0f246be6dfba2996eb90, type: 2}
