+++ UnityProject/Assets/ScriptableObjects/Speech/Hiss.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 772eef0e46246a146adb3c48809e2b4f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
